api/requirements.txt
# FastAPI
fastapi>=0.100.0
uvicorn[standard]
python-multipart

# Database
sqlalchemy>=2.0.0
psycopg2-binary
alembic

# Security
python-jose[cryptography]
passlib[bcrypt]
email-validator

# Celery
redis
celery

# HTTP Client
httpx>=0.24.0

# Vector Store
qdrant-client>=1.6.0

# Graph Store
neo4j>=5.0.0

# Configuration
pydantic-settings

# Utils
python-dotenv
